zpi/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.1
django-bootstrap5==22.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
idna==3.4
lxml==4.9.1
psycopg2-binary==2.9.3
pycparser==2.21
python-docx==0.8.11
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
